# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Budget Tracker API
  description: API for budget tracking application with versioned endpoints
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
  - url: http://localhost:5173
    description: Vite development proxy

paths:
  /api/v0/auth/status:
    get:
      tags:
        - Authentication
      summary: Get authentication status
      responses:
        '200':
          description: Authentication status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /api/v0/categories:
    get:
      tags:
        - Categories
      summary: Get categories
      parameters:
        - name: type
          in: query
          description: Filter by category type
          schema:
            type: string
            enum: [income, subscription, bill]
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /api/v0/subscriptions:
    get:
      tags:
        - Subscriptions
      summary: List subscriptions
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            default: name
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    
    post:
      tags:
        - Subscriptions
      summary: Create subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          type: object
          description: The response data
        error:
          $ref: '#/components/schemas/ErrorInfo'
        timestamp:
          type: string
          format: date-time
          description: Response timestamp

    ErrorInfo:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message

    ApiErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: false
            data:
              type: object
              nullable: true
              example: null

    AuthStatusResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            data:
              type: object
              properties:
                authenticated:
                  type: boolean
                username:
                  type: string
                  nullable: true

    CategoriesResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            data:
              type: array
              items:
                $ref: '#/components/schemas/Category'

    SubscriptionsResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            data:
              $ref: '#/components/schemas/PagedSubscriptions'

    SubscriptionResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            data:
              $ref: '#/components/schemas/Subscription'

    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [INCOME, SUBSCRIPTION, BILL]

    PagedSubscriptions:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Subscription'
        pageable:
          type: object
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    Subscription:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cost:
          type: number
          format: decimal
        currency:
          type: string
        period:
          type: string
        categoryId:
          type: integer
          format: int64

    SubscriptionRequest:
      type: object
      required:
        - name
        - cost
        - currency
        - period
        - categoryId
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        cost:
          type: number
          format: decimal
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
        period:
          type: string
          enum: [WEEKLY, MONTHLY, QUARTERLY, YEARLY]
        categoryId:
          type: integer
          format: int64

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: []